+++ be/src/agent/task_worker_pool.cpp
@@ -743,7 +743,8 @@ void* ReportResourceUsageTaskWorkerPool::_worker_thread_callback(void* arg_this)
         request.__set_report_version(g_report_version.load(std::memory_order_relaxed));
 
         TResourceUsage resource_usage;
-        resource_usage.__set_num_running_queries(ExecEnv::GetInstance()->query_context_mgr()->size());
+        resource_usage.__set_num_running_queries(
+                static_cast<int32_t>(ExecEnv::GetInstance()->query_context_mgr()->size()));
         resource_usage.__set_mem_used_bytes(ExecEnv::GetInstance()->process_mem_tracker()->consumption());
         resource_usage.__set_mem_limit_bytes(ExecEnv::GetInstance()->process_mem_tracker()->limit());
         worker_pool_this->_cpu_usage_recorder.update_interval();
