+++ be/src/exec/analytor.cpp
@@ -294,7 +294,8 @@ Status Analytor::open(RuntimeState* state) {
                 RETURN_IF_ERROR(st);
             }
         }
-        AggDataPtr agg_states = _mem_pool->allocate_aligned(_agg_states_total_size, _max_agg_state_align_size);
+        AggDataPtr agg_states =
+                _mem_pool->allocate_aligned(_agg_states_total_size, static_cast<int>(_max_agg_state_align_size));
         _managed_fn_states.emplace_back(std::make_unique<ManagedFunctionStates>(&_agg_fn_ctxs, agg_states, this));
         return Status::OK();
     };
@@ -666,13 +667,13 @@ void Analytor::remove_unused_buffer_values(RuntimeState* state) {
     _peer_group_start -= remove_count;
     _peer_group_end -= remove_count;
     _found_peer_group_end.second -= remove_count;
-    int32_t candidate_partition_end_size = _candidate_partition_ends.size();
+    int32_t candidate_partition_end_size = static_cast<int32_t>(_candidate_partition_ends.size());
     while (--candidate_partition_end_size >= 0) {
         auto peek = _candidate_partition_ends.front();
         _candidate_partition_ends.pop();
         _candidate_partition_ends.push(peek - remove_count);
     }
-    int32_t candidate_peer_group_end_size = _candidate_peer_group_ends.size();
+    int32_t candidate_peer_group_end_size = static_cast<int32_t>(_candidate_peer_group_ends.size());
     while (--candidate_peer_group_end_size >= 0) {
         auto peek = _candidate_peer_group_ends.front();
         _candidate_peer_group_ends.pop();
