+++ be/src/exec/hash_joiner.cpp
@@ -382,7 +382,7 @@ Status HashJoiner::_calc_filter_for_other_conjunct(ChunkPtr* chunk, Filter& filt
     }
 
     if (!filter_all) {
-        int zero_count = SIMD::count_zero(filter.data(), filter.size());
+        size_t zero_count = SIMD::count_zero(filter.data(), filter.size());
         if (zero_count == 0) {
             hit_all = true;
         }
